% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Book1_TheRingsetsout}
\alias{Book1_TheRingsetsout}
\title{The Ring sets out}
\format{a text file containing the text which has been scraped from htm files}
\source{
\url{http://flyingmoose.org/tolksarc/book/book.htm}
}
\usage{
Book1_TheRingsetsout
}
\description{
the six novel books of the lord of the rings
}
\details{
This dataset contains the first book of The Lord of the Rings
formatted for the convinient analysis
}
\keyword{datasets}
