% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Book2_TheRinggoessouth}
\alias{Book2_TheRinggoessouth}
\title{The Ring goes south}
\format{a text file containing the text which has been scraped from htm files}
\source{
\url{http://flyingmoose.org/tolksarc/book/book.htm}
}
\usage{
Book2_TheRinggoessouth
}
\description{
the six novel books of the lord of the rings
}
\details{
This dataset contains the second book of The Lord of the Rings
formatted for the convinient analysis
}
\keyword{datasets}
