% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Book5_TheWaroftheRing}
\alias{Book5_TheWaroftheRing}
\title{The War of the Ring}
\format{a text file containing the text which has been scraped from htm files}
\source{
\url{http://flyingmoose.org/tolksarc/book/book.htm}
}
\usage{
Book5_TheWaroftheRing
}
\description{
the six novel books of the lord of the rings
}
\details{
This dataset contains the fifth book of The Lord of the Rings
formatted for the convinient analysis
}
\keyword{datasets}
