% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Book6_TheEndoftheThirdAge}
\alias{Book6_TheEndoftheThirdAge}
\title{The End of the Third Age}
\format{a text file containing the text which has been scraped from htm files}
\source{
\url{http://flyingmoose.org/tolksarc/book/book.htm}
}
\usage{
Book6_TheEndoftheThirdAge
}
\description{
the six novel books of the lord of the rings
}
\details{
This dataset contains the sixth book of The Lord of the Rings
formatted for the convinient analysis
}
\keyword{datasets}
