% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TheLordoftheRings}
\alias{TheLordoftheRings}
\title{The Lord of the Rings}
\format{a text file containing the text which has been scraped from htm files}
\source{
\url{http://flyingmoose.org/tolksarc/book/book.htm}
}
\usage{
TheLordoftheRings
}
\description{
the six novel books of the lord of the rings
}
\details{
This package contains a complete novel set of The Lord of the Rings
The dataset below includes a complete novel set of The Lord of the Rings
formatted for the convinient analysis
}
\keyword{datasets}
